Move logical to separate file.
